only delete defs if necessary

+++ scripts/remove_word.py
@@ -38,29 +38,30 @@ if len(query1_results['results']) > 0:
     print("Word is already in use. Aborting.")
     sys.exit(1)
 
-#query2 = "select * from defs where word = '{word}'".format(word=WORD)
-#query2_results = run_sqlite_query(query2)
-#if len(query2_results['results']) > 0:
-#    print("Word is already defined. Aborting.")
-#    sys.exit(1)
+query2 = "select * from defs where word = '{word}'".format(word=WORD)
+query2_results = run_sqlite_query(query2)
+if len(query2_results['results']) > 0:
+    print("word is already defined. clearing out existing defs...")
 
-query30 = "delete from defs_log where word = '{word}'".format(word=WORD)
-query30_results = run_sqlite_query(query30)
-if not query30_results["success"]:
-    print("Failed to delete word defs_log. Aborting.")
-    sys.exit(1)
 
-query31 = "delete from defs where word = '{word}'".format(word=WORD)
-query31_results = run_sqlite_query(query31)
-if not query31_results["success"]:
-    print("Failed to delete word defs. Aborting.")
-    sys.exit(1)
+    query30 = "delete from defs_log where word = '{word}'".format(word=WORD)
+    query30_results = run_sqlite_query(query30)
+    if not query30_results["success"]:
+        print("Failed to delete word defs_log. Aborting.")
+        sys.exit(1)
 
-preview = "--preview false"
-if ENV == "dev":
-    preview = "--preview true"
-command = """npx wrangler -e {env} kv key delete {word} --binding WORDS {preview}""".format(env=ENV, word=WORD, preview=preview)
-subprocess.run(command, shell=True, text=True, check=True)
+    query31 = "delete from defs where word = '{word}'".format(word=WORD)
+    query31_results = run_sqlite_query(query31)
+    if not query31_results["success"]:
+        print("Failed to delete word defs. Aborting.")
+        sys.exit(1)
+
+    preview = "--preview false"
+    if ENV == "dev":
+        preview = "--preview true"
+
+    command = """npx wrangler -e {env} kv key delete {word} --binding WORDS {preview}""".format(env=ENV, word=WORD, preview=preview)
+    subprocess.run(command, shell=True, text=True, check=True)
 
 query3 = "delete from words where word = '{word}'".format(word=WORD)
 query3_results = run_sqlite_query(query3)
